Split LogAccount to LogAccountSelector and LogAccountCreator
