+++ src/Components/LogAccountSelector.tsx
@@ -1,18 +1,17 @@
 // react
-import React, {ChangeEvent, useCallback, useEffect, useMemo, useState} from "react";
+import React, {useCallback, useEffect, useState} from "react";
 // MUI
 import {Box, Accordion, AccordionSummary, AccordionDetails,
-        Typography, styled, Button, Divider, TextField} from '@mui/material'
+        Typography, Button} from '@mui/material'
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 // MUI styling
-import {HiddenInput, LogAccountsMUI} from "./MUIStyling/LogAccountsMUI";
+import {LogAccountsMUI} from "./MUIStyling/LogAccountsMUI";
 // global vars
 import {useGrid} from "../Providers/ProviderGrid";
 import {useTag} from "../Providers/ProviderTag";
 import {useFilter} from "../Providers/ProviderFilter";
 // types and interfaces
 import {Row} from "../Types/Row";
-import {IAccount} from "../Types/IAccount";
 import {IAccountData} from "../Types/IAccountData";
 import {floatPercentSet, integerSet, floatRoundedSet, checkboxSet, durationSet} from "../Types/colDefSets";
 import {defaultColDefs} from "../Types/defaultColDefs";
@@ -21,36 +20,22 @@ import {ColDef, ValueFormatterParams, ValueGetterParams} from "ag-grid-community
 // ag grid: custom filters
 import FilterDuration from "../FiltersComponents/FilterDuration";
 import FilterCheckboxSet from "../FiltersComponents/FilterCheckboxSet";
+import {useInitializer} from "../Providers/ProviderIniitalizer";
+import {useNavigate} from "react-router-dom"
 
-export default function LogAccounts() {
+export default function LogAccountSelector() {
+    //
+    const {accountNames} = useInitializer();
+    //
+    const nav = useNavigate();
     // global vars
-    const {gridRef, colDefs, setAccounts, accounts, setRowData, setColDefs} = useGrid();
-    const {tagDefs, setTagDefs} = useTag();
+    const {accounts, setRowData, setColDefs} = useGrid();
+    const {setTagDefs} = useTag();
     const {setUniqueAccount, setUniqueSymbol} = useFilter();
-    // MUI; hidden file upload form
-    const VisuallyHiddenInput = styled('input')(HiddenInput);
     // store raw user log data
     const [gridData, setGridData] = useState<string | null>(null);
-    // store account name
-    const [accountName, setAccountName] = useState<string>("");
-    // store filename
-    const [filename, setFilename] = useState<string>("");
     // store user log file
     const [rawString, setRawString] = useState<ArrayBuffer | string | null>(null);
-
-    // read in and save user data
-    const readInFile = useCallback((data:File | null) => {
-        if(!data || !data?.type.startsWith("text/plain")) return;
-        // assign filename received
-        setFilename(data.name);
-        // read file
-        const reader = new FileReader();
-        reader.onload = () => {
-            // save data
-            setRawString(reader.result);
-        }
-        reader.readAsBinaryString(data);
-    }, [])
 
     // initialize grid data
     useEffect(() => {
@@ -64,72 +49,6 @@ export default function LogAccounts() {
             handleColDefs(defaultColDefs, []);
         }
     }, [rawString, setGridData])
-
-    // splits sc defs from user-defined tags
-    const unsplitDef = useMemo(() => {
-        // return existing defs if no tags are found
-        if(tagDefs.length < 1) return colDefs;
-        // copy column defs
-        const tempDefs = [...gridRef?.current?.api?.getColumnDefs() as ColDef<Row>[]];
-        // filter out tag defs
-        return tempDefs.filter((i) => {
-            const idx = tagDefs.indexOf(i);
-            return idx < 0;
-        })
-    }, [colDefs, tagDefs, gridRef])
-
-    const refreshAccounts = useCallback(async() => {
-        // fetch data
-        const res = await fetch("/api/retrieveAccounts")
-        // if any error occurs
-        if(!res.ok) {
-            const err = res.text();
-            console.error("Error occurred in fetchAccounts: ", res.status, err);
-        }
-        // get values
-        const data = await res.json();
-        // if not empty, initialize account map
-        if(data?.accounts?.length > 0) {
-            const temp:Map<string, IAccountData> = new Map<string, IAccountData>();
-            data.accounts.forEach((i:IAccount) => {
-                temp.set(i.AccName, i.Data);
-            })
-            setAccounts(temp);
-        }
-    }, [setAccounts])
-
-    // save data as new account
-    const createAccount = useCallback(async () => {
-        if(!gridRef?.current) return;
-        // get row data
-        const rowData:Row[] = [];
-        gridRef.current.api?.forEachNode((i) => {
-            if(i?.data) rowData.push(i.data);
-        });
-        // initialize account type
-        const accData:IAccountData = {
-            RowData: rowData,
-            ColDefs: unsplitDef,
-            TagDefs: tagDefs
-        };
-        const acc:IAccount = {
-            AccName: accountName,
-            Data: accData
-        };
-        // send data to backend
-        const res = await fetch("api/saveNewAccount", {
-            method: "POST",
-            headers: {'Content-Type': "application/json"},
-            body: JSON.stringify({acc: acc})
-        });
-        // if any error occurs, prompt err msg to console
-        if(!res.ok) {
-            const text = await res.text();
-            console.error("Error occurred in createAccount():", res.status, text);
-        } else
-            await refreshAccounts();
-
-    }, [gridRef, unsplitDef, tagDefs, accountName, refreshAccounts])
 
     // sends raw file to backend and retrieves split array
     const toBackend = useCallback( async () => {
@@ -223,15 +142,6 @@ export default function LogAccounts() {
         setColDefs(temp);
     }, [setColDefs, setTagDefs, convertDuration])
 
-    // set selected account to grid
-    const setAccount = useCallback((acc:string) => {
-        if(!accounts.has(acc)) return;
-        const account = accounts.get(acc) as IAccountData;
-        // assign column/tag definitions and row data
-        setRowData(account.RowData);
-        handleColDefs(account.ColDefs, account.TagDefs);
-    }, [accounts, setRowData, handleColDefs])
-
     // if user uploads a file, call toBackend (send data to backend)
     useEffect(() => {
         if(!gridData) return;
@@ -242,19 +152,19 @@ export default function LogAccounts() {
         <Box sx={LogAccountsMUI.Container}>
             <Accordion defaultExpanded>
                 <AccordionSummary expandIcon={<ExpandMoreIcon/>}>
-                    <Typography>Accounts ({accounts.size})</Typography>
+                    <Typography>Accounts ({accountNames.length})</Typography>
                 </AccordionSummary>
                 <AccordionDetails>
-                    {accounts.size > 0 && Array.from(accounts.keys()).map((k:string) => (
+                    {accountNames.length > 0 && accountNames.map((i:string) => (
                         <Accordion>
                             <AccordionSummary expandIcon={<ExpandMoreIcon/>}>
-                                <Typography>{k}</Typography>
+                                <Typography>{i}</Typography>
                             </AccordionSummary>
                             <AccordionDetails>
                                 <Box sx={LogAccountsMUI.TabFlexContainer}>
                                     <Button
                                         variant="outlined"
-                                        onClick={() => setAccount(k)}
+                                        onClick={() => nav(`/${i}`)}
                                     >
                                         Load
                                     </Button>
@@ -262,55 +172,6 @@ export default function LogAccounts() {
                             </AccordionDetails>
                         </Accordion>
                     ))}
-                </AccordionDetails>
-            </Accordion>
-            <Accordion defaultExpanded>
-                <AccordionSummary expandIcon={<ExpandMoreIcon/>}>
-                    <Typography>Create new account</Typography>
-                </AccordionSummary>
-                <AccordionDetails>
-                    <Box sx={LogAccountsMUI.TabFlexCenteredContainer}>
-                        <TextField
-                            label={"Account Name"}
-                            value={accountName}
-                            onChange={(e:ChangeEvent<HTMLInputElement>) => setAccountName(e.target.value)}
-                        />
-                    </Box>
-                    <Divider sx={LogAccountsMUI.Divider}/>
-                    <Box sx={LogAccountsMUI.TabFlexContainer}>
-                        <Box sx={LogAccountsMUI.UploadGridContainer}>
-                            <Box sx={LogAccountsMUI.UploadGridCol1}>
-                                <Button
-                                    variant="outlined"
-                                    component="label"
-                                    size="small"
-                                    sx={LogAccountsMUI.UploadButton}
-                                >
-                                    Upload Log
-                                    <VisuallyHiddenInput
-                                        type="file"
-                                        onChange={(e:ChangeEvent<HTMLInputElement>) =>
-                                            readInFile(e.target.files?.[0] ?? null)}
-                                        accept="csv"
-                                    />
-                                </Button>
-                            </Box>
-                            <Box sx={LogAccountsMUI.UploadGridCol2}>
-                                    <Typography variant="subtitle2">
-                                        {rawString ? filename : "No file selected"}
-                                    </Typography>
-                            </Box>
-                        </Box>
-                    </Box>
-                    <Divider sx={LogAccountsMUI.Divider}/>
-                    <Box sx={LogAccountsMUI.TabFlexCenteredContainer}>
-                        <Button
-                            variant="outlined"
-                            disabled={!rawString || accountName === ""}
-                            onClick={createAccount}>
-                            Save Account
-                        </Button>
-                    </Box>
                 </AccordionDetails>
             </Accordion>
         </Box>
