+++ src/Components/LogLayout.tsx
@@ -8,7 +8,7 @@ import {Layout1} from '../Layouts/Layout1';
 import LogTable from "./LogTable"
 import LogColVisibility from "./LogColVisibility";
 import LogTags from "./LogTags";
-import LogAccounts from "./LogAccounts";
+import LogAccountSelector from "./LogAccountSelector";
 
 export default function LogLayout() {
     // reference for flexlayout (used in the future)
@@ -23,7 +23,7 @@ export default function LogLayout() {
         const component = node.getComponent();
         // render additional components based on value
         switch(component) {
-            case "Accounts": return <LogAccounts/>
+            case "Accounts": return <LogAccountSelector/>
             case "ColumnVisibility": return <LogColVisibility/>
             case "Tags": return <LogTags/>
             case "Table": return <LogTable/>
